split deploy components into separate files
